+++ src/app/vulnerabilities/Base Tag Hijacking.php
@@ -38,3 +38,38 @@ $app->get(
                         ->withStatus(200);
     }
 );
+
+$app->get(
+    '/base-tag-hijacking?val=nara',
+    function (Request $request, Response $response, $args) {
+        $query = $request->getQueryParams()['val'];
+        $response->getBody()->write('
+        <html>
+            <head>
+                <title>Base Tag Hijacking</title>
+                <base href="' . $query .'" target="_blank">
+                <style>
+                    body {
+                        background-color: #111;
+                    }
+                    a {
+                        text-align: center;
+                        color: white;
+                        font-family: sans-serif;
+                        text-decoration: none;
+                    }
+                    img {
+                        border: 2px solid lime;
+                    }
+                </style>
+            </head>
+            <body>
+                <img src="resources/images/goodbye.jpeg" alt="Goodbye Friend" style="display: block; margin: auto;">
+                <a href="resources/images/goodbye.jpeg">Download image from our secure website</a>
+            </body>
+        </html>'
+    );
+        return $response->withHeader("content-type", "text/html")
+                        ->withStatus(200);
+    }
+);