+++ CHANGELOG.md
@@ -39,6 +39,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Vulnerability: Laravel Ignition Reflected XSS
 - Vulnerability: S3 Bucket Publicly Accessible
 - Vulnerability: Arbitrary File Read Next.js
+- Vulnerability: Chrome Logger Information Disclosures
+- Vulnerability: Apache Tomcat Examples Directory
+- Vulnerability: Merurial Repository Found
+- Vulnerability: Flask Debug Mode
+- Vulnerability: Drupal backup_migrate
 - GIF Favicon
 - Unauthenticated Gitlab SSRF CVE 2021-22214 Demonstration
 - Software Versions List